Isolate code to create the reports (JSON and HTML)
Moved some code from 'Result' to a couple of reporters: one for JSON and
one for Handlebars). Also added an interface to make easy future
reporters.
